requirements.txt
# YOLO Training Dependencies

# Core ML framework
torch>=2.0.0
torchvision>=0.15.0

# YOLO implementation
ultralytics>=8.0.0

# Configuration and utilities
pyyaml>=6.0
numpy>=1.24.0
pillow>=9.0.0

# Logging and monitoring (optional)
tensorboard>=2.12.0
wandb>=0.15.0

# For distributed training (optional)
# pytorch-lightning>=2.0.0
